Initial work migrating to mjml email templates
