+++ src/dispatch/conversation/models.py
@@ -5,7 +5,7 @@ from typing import Optional
 from sqlalchemy import Column, String, Integer
 
 from dispatch.database import Base
-from dispatch.messaging import INCIDENT_CONVERSATION_DESCRIPTION
+from dispatch.messaging.strings import INCIDENT_CONVERSATION_DESCRIPTION
 from dispatch.models import DispatchBase, ResourceMixin
