+++ src/dispatch/document/scheduled.py
@@ -8,7 +8,7 @@ from sqlalchemy import func
 
 from dispatch.database import SessionLocal
 from dispatch.config import DISPATCH_HELP_EMAIL
-from dispatch.messaging import DOCUMENT_EVERGREEN_REMINDER
+from dispatch.messaging.strings import DOCUMENT_EVERGREEN_REMINDER
 from dispatch.decorators import background_task
 from dispatch.plugin import service as plugin_service
 from dispatch.route import service as route_service
