+++ src/dispatch/messaging/email/templates/executive_report.mjml
@@ -1,0 +1,86 @@
+{% extends 'templates/base.mjml' %}
+{% block description %}
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+<mj-column width="100%">
+    <mj-text color="#212b35" font-weight="bold" font-size="20px">
+        You're receiving this executive report, because you're a member of this incident's
+        notifications Google group ({{ notifications_group }}).
+    </mj-text>
+    <mj-text color="#212b35" font-weight="bold" font-size="20px">
+        Please be aware that this report is a point in time update and
+        the details below are likely to change as the incident evolves.
+        No direct action from you is currently required.
+    </mj-text>
+    </mj-column>
+</mj-section>
+{% endblock %}
+
+{% block items %}
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Incident Title</b></li>
+                <li>{{ title }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+<mj-section>
+    <mj-column>
+        <mj-divider border-width="1px" border-color="lightgrey" />
+    </mj-column>
+</mj-section>
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Current Status</b></li>
+                <li>{{ current_status }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+<mj-section>
+    <mj-column>
+        <mj-divider border-width="1px" border-color="lightgrey" />
+    </mj-column>
+</mj-section>
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Overview</b></li>
+                <li>{{ overview }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+<mj-section>
+    <mj-column>
+        <mj-divider border-width="1px" border-color="lightgrey" />
+    </mj-column>
+</mj-section>
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Next Steps</b></li>
+                <li>{{ next_steps }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+<mj-section>
+    <mj-column>
+        <mj-divider border-width="1px" border-color="lightgrey" />
+    </mj-column>
+</mj-section>
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+<mj-column width="100%">
+    <mj-text color="#212b35" font-weight="bold" font-size="20px">
+        A document version of the report can be found <a href="{{ weblink }}" target="_blank" rel="noopener noreferrer">here</a>. Please, use the document for questions and feedback.
+    </mj-text>
+    </mj-column>
+</mj-section>
+{% endblock %}
