+++ src/dispatch/messaging/email/templates/tactical_report.mjml
@@ -1,0 +1,73 @@
+{% extends 'templates/base.mjml' %}
+{% block description %}
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+<mj-column width="100%">
+    <mj-text color="#212b35" font-weight="bold" font-size="20px">
+        This message is for <b>notification</b> purposes only. No direct action of you
+        is currently required.
+    </mj-text>
+    <mj-text color="#212b35" font-weight="bold" font-size="20px">
+        Please be aware that this is a point in time update and the details below will
+        likely change.
+    </mj-text>
+    </mj-column>
+</mj-section>
+{% endblock %}
+
+{% block items %}
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Incident Title</b></li>
+                <li>{{ title }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+<mj-section>
+    <mj-column>
+        <mj-divider border-width="1px" border-color="lightgrey" />
+    </mj-column>
+</mj-section>
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Conditions</b></li>
+                <li>{{ conditions }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+<mj-section>
+    <mj-column>
+        <mj-divider border-width="1px" border-color="lightgrey" />
+    </mj-column>
+</mj-section>
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Actions</b></li>
+                <li>{{ actions }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+<mj-section>
+    <mj-column>
+        <mj-divider border-width="1px" border-color="lightgrey" />
+    </mj-column>
+</mj-section>
+<mj-section background-color="#ffffff" padding-left="15px" padding-right="15px">
+    <mj-column width="100%">
+        <mj-text color="#212b35" font-weight="bold" font-size="20px">
+            <ul style="list-style-type:none; padding-left: 0;">
+                <li><b>Needs</b></li>
+                <li>{{ needs }}</li>
+            </ul>
+        </mj-text>
+    </mj-column>
+</mj-section>
+{% endblock %}
