+++ src/dispatch/plugins/dispatch_google/gmail/plugin.py
@@ -9,18 +9,12 @@ from email.mime.text import MIMEText
 from typing import Dict, List, Optional
 import base64
 import logging
-import os
-import platform
 
 from tenacity import retry, stop_after_attempt
 
 from dispatch.decorators import apply, counter, timer
-from dispatch.messaging import (
-    DOCUMENT_EVERGREEN_REMINDER_DESCRIPTION,
-    INCIDENT_FEEDBACK_DAILY_DIGEST_DESCRIPTION,
-    INCIDENT_TASK_REMINDER_DESCRIPTION,
+from dispatch.messaging.strings import (
     MessageType,
-    render_message_template,
 )
 from dispatch.plugins.bases import EmailPlugin
 from dispatch.plugins.dispatch_google import gmail as google_gmail_plugin
@@ -30,7 +24,7 @@ from dispatch.plugins.dispatch_google.config import (
     GOOGLE_SERVICE_ACCOUNT_DELEGATED_ACCOUNT,
 )
 
-from .filters import env
+from dispatch.messaging.email.utils import create_message_body, create_multi_message_body
 
 
 log = logging.getLogger(__name__)
@@ -55,71 +49,6 @@ def create_html_message(recipient: str, cc: str, subject: str, body: str) -> Dic
     message["from"] = GOOGLE_SERVICE_ACCOUNT_DELEGATED_ACCOUNT
     message["subject"] = subject
     return {"raw": base64.urlsafe_b64encode(message.as_bytes()).decode()}
-
-
-def get_template(message_type: MessageType):
-    """Fetches the correct template based on the message type."""
-    template_map = {
-        MessageType.incident_executive_report: ("executive_report.html", None),
-        MessageType.incident_notification: ("notification.html", None),
-        MessageType.incident_participant_welcome: ("notification.html", None),
-        MessageType.incident_tactical_report: ("tactical_report.html", None),
-        MessageType.incident_task_reminder: (
-            "task_notification.html",
-            INCIDENT_TASK_REMINDER_DESCRIPTION,
-        ),
-        MessageType.document_evergreen_reminder: (
-            "document_evergreen_reminder.html",
-            DOCUMENT_EVERGREEN_REMINDER_DESCRIPTION,
-        ),
-        MessageType.incident_feedback_daily_digest: (
-            "feedback_notification.html",
-            INCIDENT_FEEDBACK_DAILY_DIGEST_DESCRIPTION,
-        ),
-    }
-
-    template_path, description = template_map.get(message_type, (None, None))
-
-    if not template_path:
-        raise Exception(f"Unable to determine template. MessageType: {message_type}")
-
-    return env.get_template(os.path.join("templates", template_path)), description
-
-
-def create_multi_message_body(
-    message_template: dict, message_type: MessageType, items: list, **kwargs
-):
-    """Creates a multi message message body based on message type."""
-    template, description = get_template(message_type)
-
-    master_map = []
-    for item in items:
-        master_map.append(render_message_template(message_template, **item))
-
-    kwargs.update({"items": master_map, "description": description})
-    return template.render(**kwargs)
-
-
-def create_message_body(message_template: dict, message_type: MessageType, **kwargs):
-    """Creates the correct message body based on message type."""
-    template, description = get_template(message_type)
-    rendered = render_message_template(message_template, **kwargs)
-    kwargs.update({"items": rendered, "description": description})
-    return template.render(**kwargs)
-
-
-def render_email(name, message):
-    """Helper function to preview your email."""
-    with open(name, "wb") as fp:
-        fp.write(message.encode("utf-8"))
-
-    if platform.system() == "Linux":
-        cwd = os.getcwd()
-        print(f"file:/{cwd}/{name}")
-    else:
-        print(
-            rf"/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome {name}"
-        )  # noqa: W605
 
 
 @apply(timer, exclude=["__init__"])
