--- src/dispatch/plugins/dispatch_google/gmail/templates/notification.html
@@ -1,10 +1,0 @@
-{% extends "templates/base.html" %} {% block items %} {% for item in items %}
-<ul style="list-style-type:none; padding-left: 0;">
-  <li>
-    <b
-      ><a href="{{ item.title_link }}" target="_blank" rel="noopener noreferrer">{{ item.title }}</a></b
-    >
-  </li>
-  <li>{{ item.text }}</li>
-</ul>
-{% endfor %} {% endblock %}
