+++ src/dispatch/plugins/dispatch_slack/messaging.py
@@ -8,7 +8,7 @@ import logging
 from typing import List, Optional
 from jinja2 import Template
 
-from dispatch.messaging import (
+from dispatch.messaging.strings import (
     INCIDENT_PARTICIPANT_SUGGESTED_READING_DESCRIPTION,
     INCIDENT_TASK_LIST_DESCRIPTION,
     INCIDENT_TASK_REMINDER_DESCRIPTION,
