+++ src/dispatch/report/messaging.py
@@ -6,7 +6,7 @@ from dispatch.conversation.enums import ConversationCommands
 from dispatch.database import SessionLocal, resolve_attr
 from dispatch.incident import service as incident_service
 from dispatch.incident.models import Incident
-from dispatch.messaging import (
+from dispatch.messaging.strings import (
     INCIDENT_EXECUTIVE_REPORT,
     INCIDENT_REPORT_REMINDER,
     INCIDENT_TACTICAL_REPORT,
